modify on_failure: destroy behavior for deployments
Since a deployment isn't created by the k8s connector (rather, it
expects to find it already existing on the cluster), the 'destroy'
behavior should not really delete it (among other things, this would
immediately cause an error on the next 'adjust'), but rather destroy the
pods, so that whatever is mis-configured is prevented from running,
while allowing a subsequent 'adjust' to start fresh pods with a different
configuration.

This change modifies on_failure: destroy to scale down the deployment to
zero replicas.