modify on_error:destroy behavior for deployments
connectors/kubernetes.py: a deployment handled by the connector is
'foreign' (not created by the connector), the connector cannot
auto-restore it if it has been deleted. Therefore, it should not delete
its deployment-under-control when the on_error setting is configured to
'destroy' (which is actually intended as 'destroy and re-create on the next
adjust').

This patch modifies the connector to set replicas=0, instead -
consistent with the behavior of the servoX predecessor (servo-k8s).